Soocial changes rebased ontop of the master
Soocial canges, like the search in the search in the failed queue. We
based our fork from the 2.0 branch, but at the moment the 2.0 is stale
and wasn't modified for 2 years.

Some of the changes we implemented were implemented independently and
slightly differently by the community. This commit is work and progress
and just a checkpoint to proceed with the cleanup. But all changes we've
made were cherry-picked ontop of the master branch.